improved folder structure using routes,controllers and data models
