+++ backend/app/api/routes.py
@@ -1,101 +1,17 @@
-from fastapi import APIRouter, Request, HTTPException
-from fastapi.responses import StreamingResponse
-from pydantic import BaseModel
-from app.services.llm import generate_answer
-from app.db.connect import connectDB
-from app.services.summarizer import summarize_text
-# from elevenlabs import ElevenLabs
-import uuid
-# import os
-# import psycopg2
+from fastapi import APIRouter
+from app.api.controllers import start_session, end_session, get_answer
+from app.api.models import UserData, UserResponse
 
-class UserData(BaseModel):
-    user_id: str
+chatbot_router = APIRouter()
 
-class UserResponse(BaseModel):
-    user_id: str
-    query: str
+@chatbot_router.get("/start_session")
+async def handle_request():
+    return await start_session()
 
-class textInput(BaseModel):
-    text: str
+@chatbot_router.post("/end_session")
+async def handle_request(body: UserData):
+    return await end_session(body)
 
-router = APIRouter()
-connection = connectDB()
-
-# ElevenLabs API
-# client = ElevenLabs(api_key=os.getenv('ELEVEN_API'))
-
-@router.post("/answer")
-async def get_answer(userRes: UserResponse):
-    user_id = userRes.user_id
-    query = userRes.query
-
-    try:
-        cursor = connection.cursor()
-        cursor.execute("SELECT chat_history FROM user_chats WHERE user_id = %s", (user_id,))
-        result = cursor.fetchone()
-        chat_history = result[0] if result else ""
-
-        answer = generate_answer(query, chat_history)
-        final_chat_history = summarize_text(str({
-            "user_query": query,
-            "chatbot_response": answer,
-            "chat_history": chat_history
-        }))
-
-        if result:
-            cursor.execute("UPDATE user_chats SET chat_history = %s WHERE user_id = %s", (final_chat_history, user_id))
-        else:
-            cursor.execute("INSERT INTO user_chats (user_id, chat_history) VALUES (%s, %s)", (user_id, final_chat_history))
-
-        connection.commit()
-        return {"query": query, "answer": answer}
-
-    except Exception as e:
-        print("Error during /answer:", str(e))
-        raise HTTPException(status_code=500, detail=str(e))
-
-@router.get("/start_session")
-async def start_session():
-    try:
-        user_id = str(uuid.uuid4())
-        cursor = connection.cursor()
-        cursor.execute("INSERT INTO user_chats (user_id) VALUES (%s)", (user_id,))
-        connection.commit()
-        return {"user_id": user_id}
-    except Exception as e:
-        raise HTTPException(status_code=500, detail=str(e))
-
-@router.post("/end_session")
-async def end_session(user: UserData):
-    try:
-        cursor = connection.cursor()
-        cursor.execute("DELETE FROM user_chats WHERE user_id = %s", (user.user_id,))
-        connection.commit()
-        return {"session_ended": user.user_id, "status": "success"}
-    except Exception as e:
-        raise HTTPException(status_code=500, detail=str(e))
-
-# @router.post("/speak")
-# async def generate_audio(input: textInput):
-#     try:
-#         audio = client.generate(
-#             text=input.text,
-#             voice="Alice",
-#             model="eleven_multilingual_v2"
-#         )
-
-#         audio_data = audio if isinstance(audio, bytes) else (
-#             audio.read() if hasattr(audio, 'read') else None
-#         )
-
-#         if audio_data is None:
-#             raise HTTPException(status_code=500, detail="Audio generation failed.")
-
-#         async def audio_stream():
-#             yield audio_data
-
-#         return StreamingResponse(audio_stream(), media_type="audio/mpeg")
-
-#     except Exception as e:
-#         raise HTTPException(status_code=500, detail=str(e))
+@chatbot_router.post("/answer")
+async def handle_request(body: UserResponse):
+    return await get_answer(body)
